requirements.txt
streamlit
streamlit_chat
langchain_community==0.2.6
pypdf
tiktoken
fastembed==0.3.2
chromadb
wikipedia
langgraph
langchainhub
docx2txt
duckduckgo-search
ipython
ipykernel  
sentence_transformers
jupyter
pyppeteer
numexpr
tqdm
pytest
pymupdf
flashrank  